requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-dialog==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.4
multidict==6.0.4
numpy==1.26.3
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Wand==0.6.13
yarl==1.9.4
